JWT authentication and authorization of users and goals

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require ('../models/goalModel')
+const User = require ('../models/userModel')
 const { text } = require('express')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  private
 const getGoals = asyncHandler(async (req, res)=> {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +21,8 @@ if(!req.body.text){
     throw new Error ('Please add a text field')
 }
 const goal = await Goal.create({
-    text : req.body.text
+    text : req.body.text,
+    user : req.user.id,
 })
     res.status(200).json(goal)
 })
@@ -33,6 +35,20 @@ const updateGoal = asyncHandler(async (req, res)=> {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+
+    }
+    //make sure the login user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true,})
 
@@ -47,6 +63,20 @@ const deleteGoal = asyncHandler(async (req, res)=> {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+
+    }
+    //make sure the login user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
     }
     await goal.deleteOne()
     res.status(200).json({id: req.params.id})
